requirement.txt
asgiref==3.2.10
Django==3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
